Buy Game Directly, Clean Up

+++ src/GameGate.Market/Application/Commands/SellGameByOrder/SellGameCommandHandler.cs
@@ -1,6 +1,6 @@
 ﻿using Application.Common.Interfaces;
+using Domain.Games;
 using MediatR;
-using Presentation.Controllers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace Application.Commands.SellGame;
-public class SellGameCommandHandler : IRequestHandler<SellGameCommand>
+public class SellGameCommandHandler : IRequestHandler<SellGameByOrderCommand>
 {
     private readonly IMarketRepository _marketRepository;
     private readonly IGamesHttpClient _gamesHttpClient;
@@ -17,17 +17,21 @@ public class SellGameCommandHandler : IRequestHandler<SellGameCommand>
         _marketRepository = marketRepository;
         _gamesHttpClient = gamesHttpClient;
     }
-    public async Task Handle(SellGameCommand request, CancellationToken cancellationToken)
+    public async Task Handle(SellGameByOrderCommand request, CancellationToken cancellationToken)
     {
         var order= _marketRepository.GetOrdersById(request.orderId);
         var buyer=   _marketRepository.GetBuyerById(order.BuyerId);
         var seller=   _marketRepository.GetSeller(order.SellerId);
         var game =await  _gamesHttpClient.GetGameByNameAsync(order.GameName);
-        //seller.Games.Remove(game);
-        //seller.SoldGames.Add(game);
+        if (order.GameName == null || buyer.Money <= 0 || buyer.Money < order.Bid || order.Bid < game.Price.PriceMinValue)
+        {
+            throw new Exception("Something wrong!");
+        }
+        seller.SoldGames?.Add(game);
+        seller.Games?.Remove(game);
         var gain = buyer.Money - order.Bid;
         seller.Money += gain;
-        _gamesHttpClient.DeleteGameAsync(order.GameName);
+        await _gamesHttpClient.DeleteGameAsync(order.GameName);
         _marketRepository.DeleteOrder(request.orderId);
     }
 }
